requirements.txt
gspread==6.1.4
google-auth==2.38.0
google-auth-oauthlib==1.2.1
google-api-python-client==2.149.0
python-docx==1.1.2
google-genai>=1.0.0
python-dotenv==1.0.1
